+++ Sources/Tutti/Tutorials/TutorialPage.swift
@@ -16,17 +16,20 @@ import Foundation
  a text and an image. If you want more, you can create a
  custom page type and use that instead.
  */
-open class TutorialPage {
+open class TutorialPage: Identifiable {
     
     public init(
+        id: UUID = UUID(),
         title: String,
         text: String,
         imageName: String? = nil) {
+        self.id = id
         self.title = title
         self.text = text
         self.imageName = imageName
     }
     
+    public let id: UUID
     public let title: String
     public let text: String
     public let imageName: String?